ai-assistant/requirements.txt
# SubstrateOS AI Assistant - Production Dependencies

# =============================================================================
# Web Framework
# =============================================================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# =============================================================================
# LLM Providers
# =============================================================================
openai>=1.10.0
anthropic>=0.18.0
tiktoken>=0.5.0

# =============================================================================
# RAG / Embeddings
# =============================================================================
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# =============================================================================
# Vector Store Options
# =============================================================================
pgvector>=0.2.4
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# =============================================================================
# Document Processing
# =============================================================================
unstructured>=0.12.0
python-docx>=1.1.0
pypdf>=3.17.0
markdown>=3.5.0

# =============================================================================
# Caching & Storage
# =============================================================================
redis>=5.0.0
aiofiles>=23.2.0

# =============================================================================
# Monitoring & Logging
# =============================================================================
structlog>=24.1.0
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0

# =============================================================================
# Security
# =============================================================================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0

# =============================================================================
# Utilities
# =============================================================================
python-dotenv>=1.0.0
tenacity>=8.2.0
orjson>=3.9.0
numpy>=1.26.0
